Refactor the packages to avoid "demo" references
