+++ pom.xml
@@ -9,7 +9,7 @@
 	<packaging>war</packaging>
 
 	<name>hapi-fhir-jpaserver-oauth</name>
-	<description>HAPI FHIR MySQL OAuth JPA Server - Example</description>
+	<description>HAPI FHIR OAuth JPA Server</description>
 
 	<!-- Note: HAPI projects use the "hapi-fhir" POM as their base to provide 
 		easy management. You do not need to use this in your own projects, so the 
@@ -246,7 +246,7 @@
 		<dependency>
 			<groupId>org.eclipse.jetty</groupId>
 			<artifactId>jetty-server</artifactId>
-			<version>[9.4.17.v20190418,)</version>
+			<version>${jetty_version}</version>
 			<scope>test</scope>
 		</dependency>
 		<dependency>
@@ -359,7 +359,7 @@
 				<artifactId>maven-surefire-plugin</artifactId>
 				<version>2.19.1</version>
 				<configuration>
-					<groups>ca.uhn.fhir.jpa.demo.ExampleServerIT</groups>
+					<groups>net.atos.ari.cdr.starter.TestServerIT</groups>
 					<skipTests>true</skipTests>
 				</configuration>
 			</plugin>
@@ -379,7 +379,7 @@
 							<includes>
 								<include>**/*ServerIT.java</include>
 							</includes>
-							<groups>ca.uhn.fhir.jpa.demo.IntegrationTest</groups>
+							<groups>net.atos.ari.cdr.starter.IntegrationTest</groups>
 							<skipTests>${skip.integration.tests}</skipTests>
 						</configuration>
 					</execution>
