--- src/main/java/ca/uhn/fhir/jpa/demo/elasticsearch/FhirServerConfig.java
@@ -1,109 +1,0 @@
-package ca.uhn.fhir.jpa.demo.elasticsearch;
-
-import java.util.Properties;
-import javax.persistence.EntityManagerFactory;
-import javax.sql.DataSource;
-
-import ca.uhn.fhir.jpa.config.BaseJavaConfigDstu3;
-import ca.uhn.fhir.jpa.dao.DaoConfig;
-import ca.uhn.fhir.jpa.search.ElasticsearchMappingProvider;
-import ca.uhn.fhir.jpa.util.SubscriptionsRequireManualActivationInterceptorDstu3;
-import ca.uhn.fhir.rest.server.interceptor.IServerInterceptor;
-import ca.uhn.fhir.rest.server.interceptor.ResponseHighlighterInterceptor;
-import org.apache.commons.dbcp2.BasicDataSource;
-import org.hibernate.jpa.HibernatePersistenceProvider;
-import org.hibernate.search.elasticsearch.cfg.ElasticsearchEnvironment;
-import org.springframework.beans.factory.annotation.Autowire;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.orm.jpa.JpaTransactionManager;
-import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
-
-/**
- * This is the configuration file for the example server integrating the ElasticSearch engine.
- */
-@Configuration
-@EnableTransactionManagement()
-public class FhirServerConfig extends BaseJavaConfigDstu3 {
-
-	/**
-	 * Configure FHIR properties around the the JPA server via this bean
-	 */
-	@Bean()
-	public DaoConfig daoConfig() {
-		DaoConfig retVal = new DaoConfig();
-		retVal.setAllowMultipleDelete(true);
-		return retVal;
-	}
-
-	/**
-	 * The following bean configures the database connection. The 'url' property value of "jdbc:derby:directory:jpaserver_derby_files;create=true" indicates that the server should save resources in a
-	 * directory called "jpaserver_derby_files".
-	 *
-	 * A URL to a remote database could also be placed here, along with login credentials and other properties supported by BasicDataSource.
-	 */
-	@Bean(destroyMethod = "close")
-	public DataSource dataSource() {
-		BasicDataSource retVal = new BasicDataSource();
-		retVal.setDriver(new org.apache.derby.jdbc.EmbeddedDriver());
-		retVal.setUrl("jdbc:derby:directory:target/jpaserver_derby_files;create=true");
-		retVal.setUsername("");
-		retVal.setPassword("");
-		return retVal;
-	}
-
-	@Bean()
-	public LocalContainerEntityManagerFactoryBean entityManagerFactory() {
-		LocalContainerEntityManagerFactoryBean retVal = new LocalContainerEntityManagerFactoryBean();
-		retVal.setPersistenceUnitName("HAPI_PU");
-		retVal.setDataSource(dataSource());
-		retVal.setPackagesToScan("ca.uhn.fhir.jpa.entity");
-		retVal.setPersistenceProvider(new HibernatePersistenceProvider());
-		retVal.setJpaProperties(jpaProperties());
-		return retVal;
-	}
-
-	private Properties jpaProperties() {
-		Properties extraProperties = new Properties();
-		extraProperties.put("hibernate.dialect", org.hibernate.dialect.DerbyTenSevenDialect.class.getName());
-		extraProperties.put("hibernate.format_sql", Boolean.TRUE.toString());
-		extraProperties.put("hibernate.show_sql", Boolean.FALSE.toString());
-		extraProperties.put("hibernate.hbm2ddl.auto", "update");
-		extraProperties.put("hibernate.jdbc.batch_size", "20");
-		extraProperties.put("hibernate.cache.use_query_cache", Boolean.FALSE.toString());
-		extraProperties.put("hibernate.cache.use_second_level_cache", Boolean.FALSE.toString());
-		extraProperties.put("hibernate.cache.use_structured_entries", Boolean.FALSE.toString());
-		extraProperties.put("hibernate.cache.use_minimal_puts", Boolean.FALSE.toString());
-
-		// the belowing properties are used for ElasticSearch integration
-		extraProperties.put(ElasticsearchEnvironment.ANALYSIS_DEFINITION_PROVIDER, ElasticsearchMappingProvider.class.getName());
-		extraProperties.put("hibernate.search.default.indexmanager", "elasticsearch");
-		extraProperties.put("hibernate.search.default.elasticsearch.host", "http://127.0.0.1:9200");
-		extraProperties.put("hibernate.search.default.elasticsearch.index_schema_management_strategy", "CREATE");
-		extraProperties.put("hibernate.search.default.elasticsearch.index_management_wait_timeout", "10000");
-		extraProperties.put("hibernate.search.default.elasticsearch.required_index_status", "yellow");
-		return extraProperties;
-	}
-
-	/**
-	 * This interceptor adds some pretty syntax highlighting in responses when a browser is detected
-	 */
-	@Bean(autowire = Autowire.BY_TYPE)
-	public ResponseHighlighterInterceptor responseHighlighterInterceptor() {
-		return new ResponseHighlighterInterceptor();
-	}
-
-	@Bean(autowire = Autowire.BY_TYPE)
-	public IServerInterceptor subscriptionSecurityInterceptor() {
-		return new SubscriptionsRequireManualActivationInterceptorDstu3();
-	}
-
-	@Bean()
-	public JpaTransactionManager transactionManager(EntityManagerFactory entityManagerFactory) {
-		JpaTransactionManager retVal = new JpaTransactionManager();
-		retVal.setEntityManagerFactory(entityManagerFactory);
-		return retVal;
-	}
-
-}
