+++ src/main/java/net/atos/ari/cdr/starter/JpaRestfulServer.java
@@ -1,15 +1,7 @@
 
-package ca.uhn.fhir.jpa.demo;
+package net.atos.ari.cdr.starter;
 
 import java.util.Collection;
-import java.util.List;
-
-import javax.servlet.ServletException;
-
-import org.hl7.fhir.dstu3.model.Bundle;
-import org.hl7.fhir.dstu3.model.Meta;
-import org.springframework.web.context.ContextLoaderListener;
-import org.springframework.web.context.WebApplicationContext;
 
 import ca.uhn.fhir.context.FhirContext;
 import ca.uhn.fhir.context.FhirVersionEnum;
@@ -19,16 +11,33 @@ import ca.uhn.fhir.jpa.provider.JpaConformanceProviderDstu2;
 import ca.uhn.fhir.jpa.provider.JpaSystemProviderDstu2;
 import ca.uhn.fhir.jpa.provider.dstu3.JpaConformanceProviderDstu3;
 import ca.uhn.fhir.jpa.provider.dstu3.JpaSystemProviderDstu3;
+import ca.uhn.fhir.jpa.provider.dstu3.TerminologyUploaderProviderDstu3;
+import ca.uhn.fhir.jpa.provider.r4.JpaConformanceProviderR4;
+import ca.uhn.fhir.jpa.provider.r4.JpaSystemProviderR4;
 import ca.uhn.fhir.jpa.search.DatabaseBackedPagingProvider;
+import ca.uhn.fhir.jpa.util.ResourceProviderFactory;
+// import ca.uhn.fhir.jpa.subscription.SubscriptionInterceptorLoader;
 import ca.uhn.fhir.model.dstu2.composite.MetaDt;
 import ca.uhn.fhir.narrative.DefaultThymeleafNarrativeGenerator;
 import ca.uhn.fhir.rest.api.EncodingEnum;
-import ca.uhn.fhir.rest.server.*;
+import ca.uhn.fhir.rest.server.ETagSupportEnum;
+import ca.uhn.fhir.rest.server.IResourceProvider;
+import ca.uhn.fhir.rest.server.RestfulServer;
 import ca.uhn.fhir.rest.server.interceptor.IServerInterceptor;
 
-public class JpaServerDemoDstu2 extends RestfulServer {
+import org.hl7.fhir.dstu3.model.Bundle;
+import org.hl7.fhir.dstu3.model.Meta;
+import org.springframework.web.context.ContextLoaderListener;
+import org.springframework.web.context.WebApplicationContext;
+
+import javax.servlet.ServletException;
+import java.util.List;
+
+public class JpaRestfulServer extends RestfulServer {
 
 	private static final long serialVersionUID = 1L;
+	
+	private WebApplicationContext myAppCtx;
 
 	@SuppressWarnings("unchecked")
 	@Override
@@ -41,11 +50,11 @@ public class JpaServerDemoDstu2 extends RestfulServer {
 		 *
 		 * If you want to use DSTU1 instead, change the following line, and change the 2 occurrences of dstu2 in web.xml to dstu1
 		 */
-		FhirVersionEnum fhirVersion = FhirVersionEnum.DSTU2;
+		FhirVersionEnum fhirVersion = FhirVersionEnum.DSTU3;
 		setFhirContext(new FhirContext(fhirVersion));
 
 		// Get the spring context from the web container (it's declared in web.xml)
-		WebApplicationContext myAppCtx = ContextLoaderListener.getCurrentWebApplicationContext();
+		myAppCtx = ContextLoaderListener.getCurrentWebApplicationContext();
 
 		/* 
 		 * The BaseJavaConfigDstu2.java class is a spring configuration
@@ -57,11 +66,16 @@ public class JpaServerDemoDstu2 extends RestfulServer {
 			resourceProviderBeanName = "myResourceProvidersDstu2";
 		} else if (fhirVersion == FhirVersionEnum.DSTU3) {
 			resourceProviderBeanName = "myResourceProvidersDstu3";
+		} else if (fhirVersion == FhirVersionEnum.R4) {
+			resourceProviderBeanName = "myResourceProvidersR4";
 		} else {
 			throw new IllegalStateException();
 		}
-		List<IResourceProvider> beans = myAppCtx.getBean(resourceProviderBeanName, List.class);
-		setResourceProviders(beans);
+		// List<IResourceProvider> beans = myAppCtx.getBean(resourceProviderBeanName, List.class);
+		// setResourceProviders(beans); 
+		
+		ResourceProviderFactory beans = myAppCtx.getBean(resourceProviderBeanName, ResourceProviderFactory.class);
+		registerProviders(beans.createProviders());
 		
 		/* 
 		 * The system provider implements non-resource-type methods, such as
@@ -72,10 +86,17 @@ public class JpaServerDemoDstu2 extends RestfulServer {
 			systemProvider = myAppCtx.getBean("mySystemProviderDstu2", JpaSystemProviderDstu2.class);
 		} else if (fhirVersion == FhirVersionEnum.DSTU3) {
 			systemProvider = myAppCtx.getBean("mySystemProviderDstu3", JpaSystemProviderDstu3.class);
+        } else if (fhirVersion == FhirVersionEnum.R4) {
+            systemProvider = myAppCtx.getBean("mySystemProviderR4", JpaSystemProviderR4.class);
 		} else {
 			throw new IllegalStateException();
 		}
-		setPlainProviders(systemProvider);
+		// setPlainProviders(systemProvider);
+
+        // setFhirContext(myAppCtx.getBean(FhirContext.class));
+
+        registerProvider(systemProvider);
+
 
 		/*
 		 * The conformance provider exports the supported resources, search parameters, etc for
@@ -94,6 +115,11 @@ public class JpaServerDemoDstu2 extends RestfulServer {
 			myAppCtx.getBean(DaoConfig.class));
 			confProvider.setImplementationDescription("Example Server");
 			setServerConformanceProvider(confProvider);
+        } else if (fhirVersion == FhirVersionEnum.R4) {
+            IFhirSystemDao<org.hl7.fhir.r4.model.Bundle, org.hl7.fhir.r4.model.Meta> systemDao = myAppCtx.getBean("mySystemDaoR4", IFhirSystemDao.class);
+            JpaConformanceProviderR4 confProvider = new JpaConformanceProviderR4(this, systemDao, myAppCtx.getBean(DaoConfig.class));
+            confProvider.setImplementationDescription("HAPI FHIR R4 Server");
+            setServerConformanceProvider(confProvider);
 		} else {
 			throw new IllegalStateException();
 		}
@@ -125,12 +151,29 @@ public class JpaServerDemoDstu2 extends RestfulServer {
 		setPagingProvider(myAppCtx.getBean(DatabaseBackedPagingProvider.class));
 
 		/*
+		 * Register interceptors for the server based on DaoConfig.getSupportedSubscriptionTypes()
+		 */
+		// SubscriptionInterceptorLoader subscriptionInterceptorLoader = myAppCtx.getBean(SubscriptionInterceptorLoader.class);
+		// subscriptionInterceptorLoader.registerInterceptors();
+		
+		/*
 		 * Load interceptors for the server from Spring (these are defined in FhirServerConfig.java)
 		 */
 		Collection<IServerInterceptor> interceptorBeans = myAppCtx.getBeansOfType(IServerInterceptor.class).values();
 		for (IServerInterceptor interceptor : interceptorBeans) {
 			this.registerInterceptor(interceptor);
 		}
+		
+		/*
+		 * If you are using DSTU3+, you may want to add a terminology uploader, which allows 
+		 * uploading of external terminologies such as Snomed CT. Note that this uploader
+		 * does not have any security attached (any anonymous user may use it by default)
+		 * so it is a potential security vulnerability. Consider using an AuthorizationInterceptor
+		 * with this feature.
+		 */
+		/* if (fhirVersion == FhirVersionEnum.DSTU3) {
+			 registerProvider(myAppCtx.getBean(TerminologyUploaderProviderDstu3.class));
+		} */
 	}
 
 }
